Move `types` to related with them folders
