+++ frontend/src/components/Cell/cellData.ts
