+++ frontend/src/components/Cell/corner.ts
